rename test and development build strategy to no_op strategy
load the production build strategy for staging as well as production
