+++ config/initializers/load_build_strategy.rb
@@ -1,1 +1,5 @@
-require "#{Rails.root}/lib/build_strategies/#{Rails.env}_build_strategy"
+if Rails.env.test? || Rails.env.development?
+  require 'build_strategies/no_op_build_strategy'
+else
+  require 'build_strategies/production_build_strategy'
+end
