+++ lib/build_strategies/no_op_build_strategy.rb
