--- lib/build_strategies/test_build_strategy.rb
@@ -1,10 +1,0 @@
-class BuildStrategy
-  class << self
-    def promote_build(build)
-    end
-    def merge_ref(ref)
-    end
-    def run_success_script(build)
-    end
-  end
-end
